// src/game/chess/type.ts
export enum Color {
    WHITE = "WHITE",
    BLACK = "BLACK",
}

export enum Fen {
    WHITE_PAWN = "P",
    WHITE_KNIGHT = "N",
    WHITE_BISHOP = "B",
    WHITE_ROOK = "R",
    WHITE_QUEEN = "Q",
    WHITE_KING = "K",
    BLACK_PAWN = "p",
    BLACK_KNIGHT = "n",
    BLACK_BISHOP = "b",
    BLACK_ROOK = "r",
    BLACK_QUEEN = "q",
    BLACK_KING = "k",
}

export enum Material {
    DEFAULT = "d",
    SAPPHIRE = "s",
    AMBER = "a",
    GLASS = "l",
    STEAL = "t",
    GOLD = "g",
    PAPER = "p",
    EMERALD = "e",
}

export enum Base {
    DEFAULT = "d",
    SAPPHIRE = "s",
    AMBER = "a",
    GLASS = "l",
}
// ...
export const CoordsSchema = z.object({
    x: z.number(),
    y: z.number(),
});
export type Coords = z.infer<typeof CoordsSchema>;

export const SafeSquaresSchema = z.map(z.string(), z.array(CoordsSchema));
export type SafeSquares = z.infer<typeof SafeSquaresSchema>;

export const PieceTypeSchema = z.object({
    fen: z.nativeEnum(Fen),
    material: z.nativeEnum(Material),
    base: z.nativeEnum(Base),
    hasMoved: z.boolean().optional(),
});
export type PieceType = z.infer<typeof PieceTypeSchema>;